Auto Fragment Js Added
