+++ blocks/footer/footer.js
@@ -1,5 +1,4 @@
-import { getMetadata } from '../../scripts/aem.js';
-import { loadFragment } from '../fragment/fragment.js';
+import { getMetadata, loadFragment } from '../../scripts/aem.js';
 
 /**
  * loads and decorates the footer
